Fixed typo & directory refactoring
